+++ src/org/hyperic/hq/appdef/shared/ServerValue.java
@@ -641,7 +641,7 @@ public class ServerValue extends AppdefResourceValue
     }
 
     public AppdefEntityID getEntityId() {
-        return AppdefEntityID.newServerID(getId().intValue());
+        return AppdefEntityID.newServerID(getId());
     }
 
 }
