+++ src/org/hyperic/hq/appdef/shared/ServiceValue.java
@@ -610,6 +610,6 @@ public class ServiceValue extends AppdefResourceValue
     }
 
     public AppdefEntityID getEntityId() {
-        return AppdefEntityID.newServiceID(getId().intValue());
+        return AppdefEntityID.newServiceID(getId());
     }
 }
